Update file paths and add new components
