+++ src/components/Item/texts/styles.ts
